Htmx apis app

+++ nexus/SIs/models.py
@@ -33,6 +33,11 @@ class SIRoleInfo(models.Model):
         help_text="The class that this role is for."
     )
     
+    all_sections = models.BooleanField(
+        default=False,
+        help_text="If this role is for all sections of the class."
+    )
+    
     class Meta:
         unique_together = [
             "position",
@@ -44,6 +49,8 @@ class SIRoleInfo(models.Model):
             raise ValueError("This role is not for an SI position.")
         if self.id is not None:
             old_role = SIRoleInfo.objects.get(id=self.id)
+            if old_role.assigned_class == self.assigned_class:
+                return super(SIRoleInfo, self).save(*args, **kwargs)
             SIReccuringShiftInfo.objects.filter(role=old_role).delete()
         super(SIRoleInfo, self).save(*args, **kwargs)
         class_times = ClassTimes.objects.filter(orignal_class=self.assigned_class)
